Moved chat stuff to Chat folder to make integration with master easier.
