+++ client/src/pages/Dashboard/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { Container, Row, Col, Nav } from "react-bootstrap";
 import { Link, Redirect } from "react-router-dom";
-import ChatApp from "../../components/ChatApp";
+import ChatApp from "../../components/Chat/ChatApp";
 
 import "./dashboard.css";
